removed experimenting folders
