+++ chrome/tw-exp.js
@@ -1,3 +1,49 @@
+/* global chrome, Promise, MutationSummary */
+
+var spoilersObj = {};
+var hidePref;
+
+// Promise for async get tags
+var p1 = new Promise(function(resolve, reject) {
+	// Get list of tags from persistent storage
+	chrome.storage.sync.get("allTags", function(allTags) {
+		if (!chrome.runtime.error) {
+			resolve(allTags);
+		}
+		else {
+			reject("runtime error");
+		}
+	});
+});
+
+
+// Get user preferences
+chrome.storage.sync.get("prefs", function(prefs) {
+	hidePref = prefs.prefs["hide"];
+	console.log("hide pref: " + hidePref);
+});
+
+
+// Hide page until alltags object is retrieved
+// document.documentElement.style.visibility = 'hidden';
+
+// Call rest of code when document is ready and promise has been fulfilled
+jQuery(document).ready( function($) {
+	p1.then( function(allTags) {
+		if (allTags.allTags != null) {
+			spoilersObj = allTags.allTags;
+		}
+		else {
+			spoilersObj = {};
+		}
+
+		console.log("START");
+		findStream();
+		observeBody();
+	});
+});
+
+
 // Hide all tweets that were loaded when documentready fired
 function findStream() {
 	var target = $(".stream");
@@ -10,7 +56,7 @@ function findStream() {
 		}
 
 		// All tweets on page have been checked, make document visible
-		document.documentElement.style.visibility = 'visible';
+		document.documentElement.style.visibility = '';
 
 		// Get tweets loaded on scroll
 		var tweetObs = new MutationSummary({
