--- experiment/firefox/data/fb-ff.js
@@ -1,8 +1,0 @@
-/* global self */
-var spoilersObj = {};
-
-// Get all tags json object from index.js
-self.port.on("spoilers", function(allTags) {
-	spoilersObj = allTags;
-});
-
